Moving tests to separate file
